+++ CHANGELOG.md
@@ -11,6 +11,8 @@ CHANGELOG
 * Moved PlatformLaunchValidator in Platform sub namespace  
 * Moved ToolLaunchValidator in Tool sub namespace
 * Fixed LtiServiceServer media type handling
+* Fixed LtiServiceClient grant assertion aud claim  
+* Fixed AgsClaim with not mandatory line item container url 
 * Updated LtiServiceServerRequestHandlerInterface signature
 * Updated Guzzle dependency to ^6.5 || ^7.0 
 * Updated documentation